+++ frontend/pages/withdraw/[tx].tsx
@@ -3,11 +3,10 @@ import TransferProgress from "@components/templates/transfer-progress";
 import { useRouter } from "next/router";
 import { QueryRouter } from "src/types/config";
 import { routes } from "src/constants/routes";
-import useWatchWithdraw from "src/api-query/use-watch-withdraw/useWatchWithdraw";
 import { useEffect, useState } from "react";
-import useEthMerkleProof from "src/api-query/use-eth-merkle-proof/useEthMerkleProof";
 import useRootManagerContract from "src/contracts/use-root-manager";
 import { useApprovedWithdrawalsStore } from "src/store/approved-withdraws";
+import { useEthMerkleProof, useWatchWithdraw } from "src/api-query/queries";
 
 /** Interval in ms for how often to query for deposit status */
 const QUERY_INTERVAL = 10000;
