+++ frontend/src/constants/CacheKeys.ts
@@ -1,5 +1,6 @@
 export const CacheKeys = {
     Deposit: "Cache.Deposit",
+    Withdraw: "Cache.Withdraw",
     Tokens: "Cache.Tokens",
     Wallet: "Cache.Wallet",
     CCDBalance: "Cache.CCDBalance",
